Emotion_Analysis/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.6.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
Django==3.1.4
djangorestframework==3.12.2
idna==2.10
JPype1==1.2.0
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.6.2
matplotlib==3.3.3
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.5
Pillow==8.0.1
pygame==2.0.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySocks==1.7.1
pytagcloud==0.3.5
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
scipy==1.5.4
seaborn==0.11.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tweepy==3.9.0
urllib3==1.26.2
xlrd==1.2.0
